float -> double in BaseLibS.MS

+++ BaseLibS/Ms/Ms2Lists.cs
@@ -7,16 +7,16 @@ namespace BaseLibS.Ms{
 		public int maxNumIms = 0;
 		public List<int> prevMs1IndexList = new List<int>();
 		public List<int> scansList = new List<int>();
-		public List<float> rtList = new List<float>();
+		public List<double> rtList = new List<double>();
 		public List<double> mzList = new List<double>();
 		public List<FragmentationTypeEnum> fragmentTypeList = new List<FragmentationTypeEnum>();
-		public List<float> ionInjectionTimesList = new List<float>();
-		public List<float> basepeakIntensityList = new List<float>();
-		public List<float> elapsedTimesList = new List<float>();
-		public List<float> energiesList = new List<float>();
-		public List<float> summationsList = new List<float>();
+		public List<double> ionInjectionTimesList = new List<double>();
+		public List<double> basepeakIntensityList = new List<double>();
+		public List<double> elapsedTimesList = new List<double>();
+		public List<double> energiesList = new List<double>();
+		public List<double> summationsList = new List<double>();
 		public List<double> monoisotopicMzList = new List<double>();
-		public List<float> ticList = new List<float>();
+		public List<double> ticList = new List<double>();
 		public List<bool> hasCentroidList = new List<bool>();
 		public List<bool> hasProfileList = new List<bool>();
 		public List<MassAnalyzerEnum> analyzerList = new List<MassAnalyzerEnum>();
@@ -24,12 +24,12 @@ namespace BaseLibS.Ms{
 		public List<double> maxMassList = new List<double>();
 		public List<double> isolationMzMinList = new List<double>();
 		public List<double> isolationMzMaxList = new List<double>();
-		public List<float> resolutionList = new List<float>();
+		public List<double> resolutionList = new List<double>();
 		public List<double> intenseCompFactor = new List<double>();
 		public List<double> emIntenseComp = new List<double>();
 		public List<double> rawOvFtT = new List<double>();
-		public List<float> agcList = new List<float>();
-		public List<float> agcFillList = new List<float>();
+		public List<double> agcList = new List<double>();
+		public List<double> agcFillList = new List<double>();
 		public List<int> nImsScans = new List<int>();
 	}
 }
